Move the osgjs binary decoders into a shared module
decodeVarint, deltaDecodeInPlace, dequantize and the spherical decoder are
about to have a second caller, and they are ports of viewer code that is
easy to get subtly wrong — a bad copy would be worse than an awkward
import. No behaviour change: the same model converts to the same 145
meshes and 725 accessors.

decodeNormals becomes decodeSpherical, since what it decodes is a
direction on the unit sphere rather than a normal specifically.
